Add example data (wheat-wheat) + first test (read_met)
The example data is a self-intercropping wheat (wheat-wheat) to test if the model is giving outputs close to a sole crop.
This exemple data is not suitable for model evaluation because the data are completely fabricated by myself for the sake of example. It is NOT from observations, but the values used are close to what we could expect in a real dataset.